Resolve https://github.com/UC-Davis-molecular-computing/scadnano/pull/653#discussion_r726196954
